implement telegram pooling package

+++ BottApp.Database/User/IUserRepository.cs
@@ -4,7 +4,7 @@ namespace BottApp.Database.User
 {
     public interface IUserRepository
     {
-        Task<UserModel> CreateUser(int uid, string firstName, string phone, bool isSendContact);
+        Task<UserModel> CreateUser(long uid, string firstName, string phone, bool isSendContact);
 
         Task<UserModel> GetOne(int id);
     }
